Add repository for the MeetingContribution

+++ src/main/java/chamapool/application/meetings/requests/Contribution.java
@@ -1,3 +1,3 @@
 package chamapool.application.meetings.requests;
 
-public record Contribution(Integer memberId, Integer amount) {}
+public record Contribution(Integer memberId, Double amount) {}
